1ne6ad-optimizing-fetch-user-activity-summary-performance/requirements.txt
psycopg2-binary==2.9.7
pytest==7.4.0
psutil==5.9.5
